+++ src/ui/components/layer-conversation-last-message.js
@@ -90,10 +90,6 @@ registerComponent('layer-conversation-last-message', {
      */
   },
   methods: {
-    // Lifecycle method
-    onRender() {
-      this.onRerender();
-    },
 
     // Lifecycle method
     onRerender() {