+++ src/ui/components/layer-conversation-list/layer-channel-item.js
@@ -95,10 +95,6 @@ registerComponent('layer-channel-item', {
   },
   methods: {
 
-    onRender() {
-      this.onRerender();
-    },
-
     onRerender() {
       if (this.item) this.nodes.title.innerHTML = this.item.name;
     },